refactor: 移动`AppHeader`
